fix: issues with font
