+++ package.json
@@ -57,7 +57,7 @@
     "lodash-es": "^4.17.21",
     "lucide-react": "^0.312.0",
     "nanoid": "^5.0.4",
-    "next": "^14.0.4",
+    "next": "^14.1.0",
     "next-auth": "^4.24.5",
     "next-nprogress-bar": "^2.1.2",
     "nodemailer": "^6.9.8",
