+++ src/styles/fonts.ts
@@ -1,8 +1,15 @@
-import { Instrument_Sans } from "next/font/google";
+import { Instrument_Sans, Roboto_Mono } from "next/font/google";
 
 export const instrumentSans = Instrument_Sans({
   variable: "--font-instrument-sans",
   subsets: ["latin"],
   display: "swap",
-  style: "normal",
+  adjustFontFallback: false,
+});
+
+export const robotoMono = Roboto_Mono({
+  subsets: ["latin"],
+  display: "swap",
+  variable: "--font-roboto-mono",
+  adjustFontFallback: false,
 });
